// src/common/rbac/permission.enum.ts
export enum Permission {
  USER_CREATE = 'USER_CREATE',

  PRODUCT_READ = 'PRODUCT_READ',
  PRODUCT_MANAGE = 'PRODUCT_MANAGE',

  PRICING_MANAGE = 'PRICING_MANAGE',

  ORDER_CREATE = 'ORDER_CREATE',
  ORDER_SUBMIT = 'ORDER_SUBMIT',
  ORDER_DELETE = 'ORDER_DELETE',

  ORDER_REVIEW = 'ORDER_REVIEW',
  ORDER_APPROVE = 'ORDER_APPROVE',
  ORDER_DECLINE = 'ORDER_DECLINE',
  ORDER_COMPLETE = 'ORDER_COMPLETE',

  CONSOLIDATE_CREATE = 'CONSOLIDATE_CREATE',
  CONSOLIDATE_READ_OWN = 'CONSOLIDATE_READ_OWN',

  EXPORT_READ = 'EXPORT_READ',
}